requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
pymongo
google-generativeai